move modules tests to own file
